add 4.20 - 4.22

+++ part4/blog-backend/controllers/blogs.js
@@ -33,6 +33,17 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return res.status(401).json({ error: 'user invalid' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
